moving mute button functionality to page wrapper
